+++ tasks/gulp/copy-to-destination.js
@@ -5,7 +5,6 @@ const configPaths = require('../../config/paths.json')
 const postcss = require('gulp-postcss')
 const autoprefixer = require('autoprefixer')
 const taskArguments = require('./task-arguments')
-const replace = require('gulp-replace')
 const filter = require('gulp-filter')
 const gulpif = require('gulp-if')
 
@@ -21,9 +20,6 @@ gulp.task('copy-files', () => {
     '!' + configPaths.src + '**/*.{yml,yaml}'
   ])
   .pipe(scssFiles)
-  // in dev mq show current media query. in production we block-comment it
-  .pipe(replace('//start:devonly', '/*start:devonly'))
-  .pipe(replace('//end:devonly', 'end:devonly*/'))
   .pipe(postcss([
     autoprefixer
   ], {syntax: require('postcss-scss')}))